refactor package + function
