+++ src/main/java/dev/StartupListener.java
@@ -15,11 +15,11 @@ import org.springframework.stereotype.Component;
 import dev.domain.entite.Absence;
 import dev.domain.entite.Collegue;
 import dev.domain.entite.ERole;
-import dev.domain.entite.EStatutDemandeAbsence;
-import dev.domain.entite.ETypeJourAbsence;
 import dev.domain.entite.JourAbsence;
 import dev.domain.entite.RoleCollegue;
 import dev.domain.entite.Version;
+import dev.domain.enums.EStatutDemandeAbsence;
+import dev.domain.enums.ETypeJourAbsence;
 import dev.repository.AbsenceRepo;
 import dev.repository.CollegueRepo;
 import dev.repository.JourAbsenceRepo;
