+++ src/main/java/dev/domain/entite/Absence.java
@@ -10,6 +10,9 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.ManyToOne;
 
+import dev.domain.enums.EStatutDemandeAbsence;
+import dev.domain.enums.ETypeJourAbsence;
+
 @Entity
 public class Absence {
