+++ src/main/java/dev/domain/entite/JourAbsence.java
@@ -7,6 +7,8 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
+import dev.domain.enums.ETypeJourAbsence;
+
 @Entity
 public class JourAbsence {
