+++ src/main/java/dev/domain/enums/EStatutDemandeAbsence.java
@@ -1,4 +1,4 @@
-package dev.domain.entite;
+package dev.domain.enums;
 
 public enum EStatutDemandeAbsence {
 	INITIALE, EN_ATTENTE_VALIDATION, VALIDEE, REJETEE
