+++ src/main/java/dev/domain/enums/ETypeJourAbsence.java
@@ -1,4 +1,4 @@
-package dev.domain.entite;
+package dev.domain.enums;
 
 public enum ETypeJourAbsence {
 	CONGE_PAYE, RTT, JOUR_FERIE, CONGE_SANS_SOLDE
